+++ src/main/java/sales/goods/service/GoodsService.java
@@ -7,13 +7,13 @@ import java.util.Map;
 
 public interface GoodsService {
 
-    public Good get(int id);
+    public Good get(long id);
 
     public List<Good> getAll();
 
     public Good save(Good goods);
 
-    public void delete(int id);
+    public void delete(long id);
 
     public List<Good> page(int page, int amount, Map<String, String> sort);
