+++ src/main/java/sales/payment/creaditCard/repository/CreditCardRepository.java
@@ -6,5 +6,5 @@ import sales.payment.creaditCard.domain.CreditCard;
 /**
  * Created by volodya on 29.07.15.
  */
-public interface CreditCardRepository extends JpaRepository<CreditCard, Integer> {
+public interface CreditCardRepository extends JpaRepository<CreditCard, Long> {
 }
